Introduce persistent cache support
SurePetFlap() doesn't take care of storage of the pcache, but it now
accepts a pcache argument to the constuctor.  Methods that update
the contents of the pcache first check for validity and won't attempt
to refresh them if not None unless the associated force argument is
True.

Nothing will change by default, but if the instantiator stores a
copy of inst.pcache and passes that back to the next instantiation
of the class, it'll save on some pointless calls.

Also:

* made debug a constructor argument (default False)
* added a network call logger when debug set true
* Disabled router status call because nothing useful that can't
  be got by other means is returned.