"Fix" method print_timeline() ID filtering - WARNING
With my Sure flap and account, petids and tagids are numerically
different, so this method always returned nothing.

The correct thing to do for my case is to look up the pet's tag ID
and compare with that.  This commit does just that, but this
represents a departure from earlier functionality.

THIS COMMIT MIGHT BREAK YOUR INSTALLATION.

If it does, we need to know and figure out how to check the right
values.