Refactor household_id handling: only update default_household
`household_id` is now an argument to the constructor.  This initialises
the default household, although `update_households()` will override this
if not valid.

Update methods now update only the default household rather than all of
them in order to minimise the number of requests.  This should be
satisfactory in most intances.  If the caller wants a different
household, all update_* methods now accept an `hid` argument.

This commit also adds a households property for code readability.