changed production API behaviour: update_pet_status()
A GET from the pet/x/position API endpoint used to return current
pet location but now returns a 405 Method not allowed result code.
I speculate that this is because this endpoint is intended to
*update* a pet's position with a POST call.

This commit uses the pet/x API endpoint instead, and extracts the
'position' element, which contains the same data that the old call
returned.  pet name, photo, activity and tag data etc are discarded.